✅ Update database module
